test : moveRule 테스트 내용을 Piece 테스트로 옮김
