+++ src/test/java/janggi/domain/piece/SoldierTest.java
@@ -1,17 +1,57 @@
 package janggi.domain.piece;
 
-import janggi.domain.mouveRule.SoldierMoveRule;
-import org.junit.jupiter.api.Test;
+import janggi.domain.FakeBoard;
+import janggi.domain.vo.Position;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 class SoldierTest {
-    private final Piece soldier = new Soldier(Team.HAN);
+    private final FakeBoard board = new FakeBoard();
 
-    @Test
-    void moveRule_연결_테스트() {
+    @ParameterizedTest
+    @CsvSource({
+            "CHO, 6, 4, 5, 4",
+            "HAN, 3, 4, 4, 4"
+    })
+    void 각_진영_전진_테스트(Team team, int fromRow, int fromCol, int toRow, int toCol) {
+        // given
+        Piece soldier = new Soldier(team);
+        Position from = new Position(fromRow, fromCol);
+        Position to = new Position(toRow, toCol);
+
         // when, then
-        assertThat(soldier.moveRule()).isInstanceOf(SoldierMoveRule.class);
+        assertThat(soldier.moveRule().canMove(from, to, board)).isTrue();
     }
 
+    @ParameterizedTest
+    @CsvSource({
+            "CHO, 6, 4, 7, 4",
+            "HAN, 3, 4, 2, 4"
+    })
+    void 각_진영_후진_불가_테스트(Team team, int fromRow, int fromCol, int toRow, int toCol) {
+        // given
+        Piece soldier = new Soldier(team);
+        Position from = new Position(fromRow, fromCol);
+        Position to = new Position(toRow, toCol);
+
+        // when, then
+        assertThat(soldier.moveRule().canMove(from, to, board)).isFalse();
+    }
+
+    @ParameterizedTest
+    @CsvSource({
+            "CHO, 5, 4, 5, 3", // 초 서쪽 이동
+            "CHO, 5, 4, 5, 5", // 초 동쪽 이동
+            "HAN, 4, 4, 4, 3", // 한 서쪽 이동
+            "HAN, 4, 4, 4, 5"  // 한 동쪽 이동
+    })
+    void 각_진영_좌우_이동_테스트(Team team, int fromRow, int fromCol, int toRow, int toCol) {
+        Piece soldier = new Soldier(team);
+        Position from = new Position(fromRow, fromCol);
+        Position to = new Position(toRow, toCol);
+
+        assertThat(soldier.moveRule().canMove(from, to, board)).isTrue();
+    }
 }
